requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==9.5.0
numpy==1.21.6
scikit-image==0.19.3
scikit-learn==1.0.2
matplotlib==3.5.3
reportlab==4.0.4
jinja2==3.1.2
aiofiles==23.2.1
pydantic==1.10.13
starlette==0.27.0
requests==2.31.0
urllib3==1.26.18

